% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_present.R
\name{all_present}
\alias{all_present}
\title{Check if all values are present in set}
\usage{
all_present(..., set = c())
}
\arguments{
\item{...}{Values to check}

\item{set}{Vector containing values}
}
\value{
TRUE or FALSE if all values are present in set
}
\description{
Check if all values are present in set
}
