% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_contributions.R
\name{analyze_contributions}
\alias{analyze_contributions}
\title{Add meta data to contributions file}
\usage{
analyze_contributions(contributions_fp, mappingfile_fp,
  metadata = c("Description"))
}
\arguments{
\item{contributions_fp}{File location of the metagenomic_contributions.py output file.}

\item{metadata}{a vector of metadata identfiers which you would like to add as columns.}

\item{mappingfile}{File location of your sample metadata.}
}
\value{
A dataframe with taxa information and sample metadata for each observed OTU.
}
\description{
This function loads a file as a matrix. It assumes that the first column
contains the rownames and the subsequent columns are the sample identifiers.
Any rows with duplicated row names will be dropped with the first one being
kepted.
}

