% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bf_ratio.R
\name{bf_ratio}
\alias{bf_ratio}
\title{Computer Bacteroidetes/Firmicutes ratio for each sample.}
\usage{
bf_ratio(phylo)
}
\arguments{
\item{phylo}{An input phyloseq object.}
}
\value{
A phyloseq object with 'log2_bf_ratio' as a new metadata column
}
\description{
This function will computer the B/F ratio based on the relative abundances
between the Bacteroidetes and Firmicutes phyla.
}
