% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_number.R
\name{bin_number}
\alias{bin_number}
\title{Calculate bin number for ggplot2}
\usage{
bin_number(data_vector, method = "fd")
}
\arguments{
\item{data_vector}{Column of dataframe being plotted}

\item{method}{`c("fd")` String indicating method to use. Currently
only the Freedman-Diaconis rule is implemented}
}
\value{
number_bins, integer containing the ideal number of bins
 for histogram
}
\description{
Calculate bin number for ggplot2
}
