\name{btools-package}
\alias{btools-package}
\alias{btools}
\docType{package}
\title{
\packageTitle{btools}
}
\description{
\packageDescription{btools}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{btools}

Maintainer: \packageMaintainer{btools}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = btools)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{btools}}
%\section{Documentation Index}{\packageIndices{btools}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
