% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/btools.r
\docType{package}
\name{btools}
\alias{btools}
\alias{btools-package}
\title{btools.}
\description{
Tools that I use regularly.
Also, includes frequently needed data.
}
\section{Overview. Includes data files}{
 recessions, stcodes, spop.a, spop.q, spi.a, spi.q.
}

\section{Examples. ht(popst)}{

}

