% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html-tools.R
\name{col_text}
\alias{col_text}
\title{Add HTML Colour to Text}
\usage{
col_text(x, colour = "red")
}
\arguments{
\item{x}{String to add flags around.}

\item{colour}{CSS colouring to be added.}
}
\value{
String.
}
\description{
Adds \verb{<span style="color: colour">x</span>} flags to add colour in html and
markdown formatting.
}
