% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_beta_diversity.R
\name{compare_beta_diversity}
\alias{compare_beta_diversity}
\title{Calculate alpha diversity statistics}
\usage{
compare_beta_diversity(phylo, x = "Day", group = "Treatment",
  test = c("adonis", "anosim"), bdiv = c("unweighted", "weighted"),
  write = F, filename = "results", ...)
}
\arguments{
\item{phylo}{A phyloseq object}

\item{x}{The variable that describes Time}

\item{group}{The variable that describes 2 or more groups}

\item{test}{The PERMANOVA test to use.}

\item{bdiv}{Beta diversity metric to calculate significance.}

\item{write}{Should results be written to a file in the current working directory.}

\item{filename}{The name of the output file if write is True.}
}
\value{
A dataframe for an PERMANOVA test over each timepoint from each two group comparison.
}
\description{
Compute p-values and multiple comparisons adjusted q-values for
two-group comparisons across multiple timepoints.
}

