% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcloneSMB.R
\name{copyto_rclone_smb}
\alias{copyto_rclone_smb}
\title{Copy file from source to destination}
\usage{
copyto_rclone_smb(
  host,
  user,
  src,
  dest,
  src_smb = TRUE,
  dest_smb = TRUE,
  password = NULL,
  rclone_location = Sys.getenv("rclone_location"),
  ...
)
}
\arguments{
\item{host}{SMB server hostname to connect to}

\item{user}{SMB username}

\item{src}{Source file/folder path}

\item{dest}{Destination file/folder path}

\item{src_smb}{Source is on smb. If false path is treated as local}

\item{dest_smb}{Destination is on smb. If false path is treated as local}

\item{password}{Plain text password. If you have an obscured password
pass the obscured password in variable \code{pass}.}

\item{rclone_location}{Location of \code{rclone}}

\item{...}{Additional variables for \link{rclone_smb_connection_string}}
}
\value{
Silent
}
\description{
Overwrite file/folder at destination with source
}
