% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_phylo.R
\name{create_phylo}
\alias{create_phylo}
\title{Generic function for creating a phyloseq object from QIIME intput data.}
\usage{
create_phylo(biom_fp, mappingfile_fp, tree_fp)
}
\arguments{
\item{biom_fp}{File location of OTU table in .biom format.}

\item{mappingfile_fp}{File location of the input sample metadata matching the input OTU table.}

\item{tree_fp}{File location of the input sample metadata matching the input OTU table.}
}
\value{
A phyloseq object to be used downstream for many different analyses.
}
\description{
This function loads an TU table (.biom format) a phylogenetic tree file (.tre) and
a sample mapping file (.txt/.tsv) to create a standard phyloseq object. It is important
to first convert the OTU table to JSON format, due to an issue currently with the biom package.
}

