% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDirectorySMB.R
\name{createDirectorySMB}
\alias{createDirectorySMB}
\title{createDirectorySMB}
\usage{
createDirectorySMB(
  username,
  password,
  my_name = Sys.info()[["nodename"]],
  remote_name,
  service_name,
  path,
  domain,
  port = 445,
  timeout = 30,
  python_location
)
}
\arguments{
\item{username}{Your SMB user name}

\item{password}{Your SMB user password, do not store this in your code!}

\item{my_name}{The machine you are running this operation on (Default : \code{Sys.info()[["nodename"]]})}

\item{remote_name}{The smb server name}

\item{service_name}{The service_name name i.e. \code{'service_name'}}

\item{path}{The SMB folder path ro create e.g. \code{'Files/Reporting/'}}

\item{domain}{SMB server domain}

\item{port}{Common port  (Default : \code{445})}

\item{timeout}{How long to keep the connection open before timeout (Default : \code{30})}

\item{python_location}{The location of your python installation}
}
\value{
Retrieve a directory listing of files/folders at \code{path}
}
\description{
Use python SMBConnection to create a directory at \code{path}
}
