% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cton}
\alias{cton}
\title{Convert character to numeric}
\usage{
cton(cvar)
}
\arguments{
\item{cvar}{The character string input. No default.}
}
\description{
\code{cton} converts character to numeric
}
\details{
Replaces spaces, comma, $, and percent sign in a string with NULL and then converts to numeric.
Keeps letters so that scientific notation will be evaluated properly.
}
\examples{
char <- "$198,234.75"
cton(char)
}
\keyword{cton}

