% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_uncertainty.R
\name{display_uncertainty}
\alias{display_uncertainty}
\title{Display Uncertainty}
\usage{
display_uncertainty(
  data,
  value = NULL,
  uncertainty = NULL,
  relative_uncertainty = NULL,
  sig_fig = 2,
  display = NULL,
  latex = FALSE,
  add_spaces = FALSE
)
}
\arguments{
\item{data}{dataframe containing values to round given an uncertainty}

\item{value}{string, variable name of values}

\item{uncertainty}{string, variable name of standard uncertainty}

\item{relative_uncertainty}{string, variable name of standard relative uncertainty}

\item{sig_fig}{int, number of significant digits should remain for the mantissa of the uncertainty, defaults to 2}

\item{display}{string, name of output column containing the data, defaults to "display"}

\item{latex}{boolean, output the display as latex instead of plain text, defaults to false}

\item{add_spaces}{boolean, add spaces to improve readability, defaults to false}
}
\value{
Original dataframe with new column containing the properly displayed value and uncertainty
}
\description{
Display Uncertainty
}
