% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_comparison.R
\name{diversity_comparison}
\alias{diversity_comparison}
\title{Compute pairwise diversity comparisons of different beta diversity metrics.}
\usage{
diversity_comparison(phylo, distance = c("jsd", "unifrac", "wunifrac",
  "dpcoa", "jaccard", "manhattan", "bray", "morisita", "horn", "chao",
  "canberra", "euclidean"))
}
\arguments{
\item{phylo}{An input phyloseq object.}

\item{distance}{Distance metrics to use for the pairwise comparisons.}
}
\value{
A dataframe of distance comparions.
}
\description{
The function will generate the distance comparisons of various estimates found in the vegan package.
The output will be in a dataframe format to generate additionally analyses of the distance
comparisons.
}

