% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dmonths}
\alias{dmonths}
\title{Add dmonths (delta-months) to a date}
\usage{
dmonths(dt, dm)
}
\arguments{
\item{dt}{The date input. No default.}

\item{dm}{The number of months to add. Can be negative. No default.}
}
\description{
\code{dmonths} add dmonths (delta-months) to a date
}
\details{
You must ensure that the inputs are acceptable .
}
\examples{
dmonths(as.Date("2012-05-01"), 8)
}
\keyword{dmonths}

