% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btools_utilities.r
\name{dollar_d}
\alias{dollar_d}
\title{Format a numeric vector in dollar (or comma) format}
\usage{
dollar_d(x, d = 0, parens = FALSE, dsign = TRUE)
}
\arguments{
\item{x}{Numeric vector.}

\item{d}{Integer, number of decimal places.}

\item{parens}{Logical - enclose negative numbers with parentheses?}

\item{dsign}{Logical - leading dollar sign?}
}
\description{
Format a numeric vector in dollar (or comma) format
}
\examples{
x <- c(-10, -1.235, -23456.789, 1, 10, 100, 1000000, 1789456.23456789)
dollar_d(x)
dollar_d(x, 1)
dollar_d(x, 1, dsign=FALSE)
dollar_d(x, 1, parens=TRUE)
dollar_d(x, 1, parens=TRUE, dsign=FALSE)
}
