% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_pd.R
\name{estimate_pd}
\alias{estimate_pd}
\title{Estimate phylogenetic diversity from phyloseq object}
\usage{
estimate_pd(phylo)
}
\arguments{
\item{phylo}{A phyloseq object with an OTU table and phylogenetic tree slot.}
}
\value{
A data.frame of phylogenetic diversity metrics.
}
\description{
Estimate the Faiths phylogenetic diverstiy from an OTU table and phylogenetic tree
}

