% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btools_utilities.r
\name{even}
\alias{even}
\title{return even elements of a vector}
\usage{
even(x)
}
\arguments{
\item{x}{vector}
}
\value{
numeric vector
}
\description{
return even elements of a vector
}
\examples{
even(1:10)
}
