% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btools_formatting.r
\name{f_dollar}
\alias{f_dollar}
\title{Format numeric vector in dollar format}
\usage{
f_dollar(num, ...)
}
\arguments{
\item{num}{numeric vector to format}

\item{...}{arguments passed to scales::label_dollar}
}
\value{
formatted values as a character vector.
}
\description{
\code{f_dollar} dollar format - calls scales::label_dollar
}
\details{
format the vector using scales::label_dollar
}
\examples{
f_dollar(1.2345)
f_dollar(1.2345, scale=1000)
f_dollar(1.2345, scale=1000, accuracy=0.01)
}
\keyword{f_dollar}
