% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btools_formatting.r
\name{f_pct}
\alias{f_pct}
\title{Format numeric vector as percent}
\usage{
f_pct(num, ...)
}
\arguments{
\item{num}{numeric vector to format}

\item{...}{arguments passed to scales::label_percent}
}
\value{
formatted values as a character vector.
}
\description{
\code{f_pct} Percent format - calls scales::label_percent
}
\details{
format the vector using scales::label_percent
}
\examples{
f_pct(1.2345)
f_pct(1.2345, scale=.1)
f_pct(1.2345, scale=.1, accuracy=0.01)
}
\keyword{f_pct}
