% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btools_dates.r
\name{fdoq}
\alias{fdoq}
\title{First day of quarter, given a date as string or date}
\usage{
fdoq(date)
}
\arguments{
\item{date}{a vector of dates as strings or dates. No default.}
}
\description{
\code{fdoq} create a date for start of quarter, from quarter, year
}
\examples{
x <- c('2010-01-03', '2011-03-03', '2011-04-07')
fdoq(x)
}
