% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btools_utilities.r
\name{fton}
\alias{fton}
\title{Factor to numeric}
\usage{
fton(fctr)
}
\arguments{
\item{fctr}{factor that we want to convert to numeric}
}
\value{
numeric vector
}
\description{
Returns a numeric vector, converted from factor.
}
\details{
Returns a pure numeric vector
}
\examples{
set.seed(1234)
fctr <- factor(sample(1:4, 50, replace=TRUE), levels=1:4)
fctr
fton(fctr)
}
\keyword{fton}
