% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAttributesSMB.R
\name{getAttributesSMB}
\alias{getAttributesSMB}
\title{getAttributesSMB}
\usage{
getAttributesSMB(
  username,
  password,
  my_name = Sys.info()[["nodename"]],
  remote_name,
  service_name,
  path,
  domain,
  port = 445,
  timeout = 30,
  python_location
)
}
\arguments{
\item{username}{Your SMB user name}

\item{password}{Your SMB user password, do not store this in your code!}

\item{my_name}{The machine you are running this operation on (Default : \code{Sys.info()[["nodename"]]})}

\item{remote_name}{The smb server name}

\item{service_name}{The service_name name i.e. \code{'service_name'}}

\item{path}{The SMB folder path where your file is e.g. \code{'Files/Reporting/Test.csv'}}

\item{domain}{SMB server domain}

\item{port}{Common port  (Default : \code{445})}

\item{timeout}{How long to keep the connection open before timeout (Default : \code{30})}

\item{python_location}{The location of your python installation}
}
\value{
Retrieve a dataframe of attributes
}
\description{
Use python SMBConnection to retrieve information about the file at \code{path} on the \code{service_name.}
}
