% Generated by roxygen2 (4.0.0): do not edit by hand
\name{getstname}
\alias{getstname}
\title{Get state name from state abbreviation}
\usage{
getstname(st)
}
\arguments{
  \item{st}{The state abbreviation.}
}
\description{
\code{getstname} get state name from state abbreviation
}
\details{
get state name from state abbreviation
}
\examples{
getstname("CO")
}
\keyword{getstname}

