% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grapes-allin-grapes.R
\name{\%allin\%}
\alias{\%allin\%}
\title{Check if all values are present in set}
\usage{
x \%allin\% y
}
\arguments{
\item{x}{Values to check}

\item{y}{Vector containing values}
}
\value{
TRUE or FALSE if all values are present in set
}
\description{
Check if all values are present in set
}
