% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight_diff.R
\name{highlight_diff}
\alias{highlight_diff}
\title{Highlight Difference Between Two Strings}
\usage{
highlight_diff(original, different, colour = "red", split = FALSE)
}
\arguments{
\item{original}{String to be compared against.}

\item{different}{New string that will be returned and higlighted.}

\item{colour}{Colour that differences will be highlighted with.}

\item{split}{Logical, whether to returne a vector of individual characters or
a collapsed single string.}
}
\value{
String vecctor.
}
\description{
Finds and highlights with html colour tags, the difference between two
strings.
}
\examples{
orig <- "AAACUCAGUU"
mut <- "AAACUUAGUU"
highlight_diff(orig, mut)
}
