% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_rcc.R
\name{import_rcc}
\alias{import_rcc}
\title{Import .RCC files}
\usage{
import_rcc(rccPath = "", useFolderPath = T)
}
\arguments{
\item{rccPath}{.RCC files from NanoString analysis. Adapted from NanoStringNorm.}

\item{useFolderPath}{Is the input rccPath a real folder, or from a dataframe.}
}
\value{
A dataframe of count data with columns as samples and rows as genes.
}
\description{
Import nanostring RCC raw files and generate a table based on sampleID's from each file.
}

