% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_pmder.R
\name{install_pmder}
\alias{install_pmder}
\title{Install pmder package}
\usage{
install_pmder()
}
\description{
Execution of this function will install the latest version of the `pmder` package into the system.
}
