% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_tools.R
\name{install_tools}
\alias{install_tools}
\title{Installs bioinformatics tools.}
\usage{
install_tools(whitelist = NA, blacklist = NA)
}
\arguments{
\item{whitelist}{Installs only whitelisted tools}

\item{blacklist}{Installs all tools available except blacklisted tools}
}
\description{
This function downloads and compiles the source files of all the
requested tools. However, it may not provide
all the libraries and dependencies necessary for their succesful installation,
therefore, they have to be installed beforehand as described in the README.md
or at https://github.com/TearsWillFall/BTools
}
