% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverson.R
\name{inverson}
\alias{inverson}
\title{Inverse Iverson Bracket}
\usage{
inverson(...)
}
\arguments{
\item{...}{Vector of 0 or 1}
}
\value{
Logical vector
}
\description{
Return FALSE for 0 and TRUE for 1
}
\examples{

`inverson(0, 1, 0, 1)`
## c(FALSE, TRUE, FALSE, TRUE)
}
