% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btools_utilities.r
\name{is.odd}
\alias{is.odd}
\title{which elements of a vector are odd}
\usage{
is.odd(x)
}
\arguments{
\item{x}{vector}
}
\value{
logical vector
}
\description{
which elements of a vector are odd
}
\examples{
is.odd(1:10)
}
