% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btools_utilities.r
\name{is_even}
\alias{is_even}
\title{which elements of a vector are even}
\usage{
is_even(x)
}
\arguments{
\item{x}{vector.}
}
\value{
logical vector
}
\description{
which elements of a vector are even
}
\examples{
is_even(1:10)
}
