% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btools_utilities.r
\name{is_true}
\alias{is_true}
\title{function to deal with NA logical values}
\usage{
is_true(x)
}
\arguments{
\item{x}{vector.}
}
\value{
logical vector
}
\description{
function to deal with NA logical values
}
