% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iverson.R
\name{iverson}
\alias{iverson}
\title{Iverson Bracket}
\usage{
iverson(...)
}
\arguments{
\item{...}{Vector of boolean}
}
\value{
Vector
}
\description{
Return 0 for FALSE and 1 for TRUE
}
\examples{
# iverson(TRUE, FALSE, TRUE, FALSE)
## c(1, 0, 1, 0)
}
