% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jags_run.R
\name{jags_run}
\alias{jags_run}
\title{Interface with JAGS}
\usage{
jags_run(
  ...,
  model = NULL,
  data = list(),
  variables = c(),
  burn_in = 1,
  steps = 1,
  thin = 1,
  directory = "models/jags/"
)
}
\arguments{
\item{...}{Collection of strings containing the JAGS model}

\item{model}{String vector containing the model parameters}

\item{data}{List, data for likelihood and prior distribution}

\item{variables}{String vector, variables to keep track of}

\item{burn_in}{Integer, number of burn in steps to perform}

\item{steps}{Number of sampling steps to perform}

\item{thin}{Thinning interval}
}
\value{
model_output Named list containing arrays of sampled posterior distributions
}
\description{
This function allows for interfacing with the JAGS (just another gibbs
sampler) for Markov Chain Monte Carlo sampling of posterior distributions
}
