% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btools_dplyrtools.r
\name{lcnames}
\alias{lcnames}
\title{Convert names of a data frame to lower case}
\usage{
lcnames(df)
}
\arguments{
\item{df}{Data frame.}
}
\value{
Data frame with lower-case names
}
\description{
Convert names of a data frame to lower case, return as data frame. This works
with the pipe operator introduced in R 4.1.
}
\examples{
library(dplyr)
df <- tibble(YEAR=2000:2010, X=10:20, y=30:40)
df
df |>
  lcnames()
}
\keyword{lcnames}
