% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcloneSMB.R
\name{ls_rclone_smb}
\alias{ls_rclone_smb}
\title{List files and folders}
\usage{
ls_rclone_smb(
  host,
  user,
  path,
  password = NULL,
  rclone_location = Sys.getenv("rclone_location"),
  files_only = FALSE,
  dirs_only = FALSE,
  recursive = FALSE,
  ...
)
}
\arguments{
\item{host}{SMB server hostname to connect to.}

\item{user}{SMB username.}

\item{path}{Folder path}

\item{password}{Plain text password. If you have an obscured password
pass the obscured password in variable \code{pass}.}

\item{rclone_location}{Location of \code{rclone}}

\item{files_only}{Do not return directories}

\item{dirs_only}{Do not return files}

\item{recursive}{Recurse into the listing}

\item{...}{Additional variables for \link{rclone_smb_connection_string}}
}
\value{
dataframe of files and folder within \code{path}
}
\description{
List files and folders
}
