% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btools_utilities.r
\name{ma}
\alias{ma}
\title{Get trailing moving average}
\usage{
ma(x, n)
}
\arguments{
\item{x}{The vector to operate on.}

\item{n}{The period of the moving average.}
}
\description{
\code{ma} Get trailing moving average
}
\details{
Moving average of the vector x.
}
\examples{
ma(7:21, 3)
}
\keyword{ma}
