% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcnparse.r
\name{mcnparse}
\alias{mcnparse}
\title{Title}
\usage{
mcnparse(deck_list, directory = "", extension = ".mcnpout", write_files = TRUE)
}
\arguments{
\item{deck_list}{String vector containing the file_names (without extension) to be used}

\item{directory}{String containing the path to the directory of the output files. By default, this will be your current working directory. The command setwd("path/to/directory") will change the working directory in R.}

\item{extension}{String containing the file_name extension for all output files if no argument is provided, this will use ".mcnpout".}

\item{write_files}{Boolean. When set to true, the data for the runs will be saved to "mcnpData.csv" and the statistical checks will be saved to "statisticalChecks.csv".}
}
\value{
named list with element data: dataframe containing the data check: dataframe containing the statistical checks
}
\description{
The data for the mcnp runs contains four columns:
}
\examples{

`mcnparse(c("run1", "run2", "run3"), directory = "", extension = ".o", write_files = FALSE)`
## Will parse the files "run1.o", "run2.o", "run3.o" and will not save the output to excel.
}
