% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcnparse.r
\name{mcnparse_identify_tally}
\alias{mcnparse_identify_tally}
\title{Identify tally locations for MCNP output}
\usage{
mcnparse_identify_tally(mcnp_output)
}
\arguments{
\item{mcnp_output}{String vector containing MCNP output file}
}
\value{
Dataframe containing each tally identifier, type, and location of the tally, the values, and number of energy bins
}
\description{
Function acts as subroutine for the mcnparse function.
}
