% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcnparse.r
\name{mcnparse_one}
\alias{mcnparse_one}
\title{Parse single MCNP output file}
\usage{
mcnparse_one(file_name, directory = "", extension = ".mcnpout")
}
\arguments{
\item{file_name}{String with name of MCNP output file}

\item{directory}{String containing the path to the directory of the output files. By default, this will be your current working directory. The command setwd("path/to/directory") will change the working directory in R.}

\item{extension}{Boolean. When set to true, the data for the runs will be saved to "mcnpData.csv" and the statistical checks will be saved to "statisticalChecks.csv".

This function will pull the data from each tally and return a dataframe containing vectors for the information.
Function acts as subroutine for the mcnparse function.}
}
\value{
Returns dataframe containing results of individual run
}
\description{
Parse single MCNP output file
}
