% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcnparse.r
\name{mcnparse_pull_tally_data}
\alias{mcnparse_pull_tally_data}
\title{Pull Tally Location Information}
\usage{
mcnparse_pull_tally_data(in_tally_lines, tally_name)
}
\arguments{
\item{in_tally_lines}{Subset of MCNP output containing tally information}

\item{tally_name}{Tally identifier}
}
\value{
Dataframe containing energy bin, tally value, relative
 error, and tally name
}
\description{
This function will pull the data from each tally and return a dataframe containing vectors for the information.
Function acts as subroutine for the mcnparse function.
}
