% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcnparse.r
\name{mcnparse_stat_results}
\alias{mcnparse_stat_results}
\title{Parse single statistical checks for MCNP output}
\usage{
mcnparse_stat_results(mcnp_output)
}
\arguments{
\item{mcnp_output}{}
}
\value{
Dataframe containing the results of MCNP statistical checks for each tally
}
\description{
Get results of statistical checks for a particular run.
Function acts as subroutine for the mcnparse function.
}
