% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mdy}
\alias{mdy}
\title{Create a date from month, day, year}
\usage{
mdy(m, d, y)
}
\arguments{
\item{m}{the month input. No default.}

\item{d}{the day input. No default.}

\item{y}{the year input. No default.}
}
\description{
\code{mdy} create a date from month, day, year
}
\details{
You must ensure that the inputs are acceptable (e.g., month is in 1:12). The function does not check.
}
\examples{
mdy(10,1,1988)
}
\keyword{mdy}

