% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btools_dates.r
\name{mdy_fn}
\alias{mdy_fn}
\title{Create a date from month, day, year}
\usage{
mdy_fn(m, d, y)
}
\arguments{
\item{m}{the month input. No default.}

\item{d}{the day input. No default.}

\item{y}{the year input. No default.}
}
\description{
\code{mdy_fn} create a date from month, day, year
}
\details{
You must ensure that the inputs are acceptable (e.g., month is in 1:12). The function does not check.
}
\examples{
mdy_fn(10, 1, 1988)
}
\keyword{mdy_fn}
