% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btools_dates.r
\name{mdyfn}
\alias{mdyfn}
\title{Create a date from month, day, year}
\usage{
mdyfn(m, d, y)
}
\arguments{
\item{m}{the month input. No default.}

\item{d}{the day input. No default.}

\item{y}{the year input. No default.}
}
\description{
\code{mdyfn} create a date from month, day, year
}
\details{
You must ensure that the inputs are acceptable (e.g., month is in 1:12). The function does not check.
}
\examples{
mdyfn(10, 1, 1988)
}
\keyword{mdyfn}
