% Generated by roxygen2 (4.0.0): do not edit by hand
\name{memory}
\alias{memory}
\title{Describe memory usage and collect garbage}
\usage{
memory()
}
\arguments{
  \item{maxnobjs}{The number of objects to display. Default
  is 5.}
}
\description{
\code{memory} converts character to numeric
}
\details{
Describes memory usage and collects garbage
}
\examples{
memory(4)
}
\keyword{memory}

