% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcloneSMB.R
\name{mkdir_rclone_smb}
\alias{mkdir_rclone_smb}
\title{Make folder}
\usage{
mkdir_rclone_smb(
  host,
  user,
  path,
  password = NULL,
  rclone_location = Sys.getenv("rclone_location"),
  ...
)
}
\arguments{
\item{host}{SMB server hostname to connect to.}

\item{user}{SMB username.}

\item{path}{SMB folder to create}

\item{password}{Plain text password. If you have an obscured password
pass the obscured password in variable \code{pass}.}

\item{rclone_location}{Location of \code{rclone}}

\item{...}{Additional variables for \link{rclone_smb_connection_string}}
}
\value{
Silent
}
\description{
Make folder
}
