% Generated by roxygen2 (4.0.1): do not edit by hand
\name{month}
\alias{month}
\title{Extract month from a date}
\usage{
month(dt)
}
\arguments{
\item{dt}{The date input. No default.}
}
\description{
\code{month} extract month from a date
}
\details{
extract month from a date
}
\examples{
month("2012-05-01")
}
\keyword{month}

