% Generated by roxygen2 (4.0.2): do not edit by hand
\name{naz}
\alias{naz}
\title{Convert NA to zero}
\usage{
naz(vec)
}
\arguments{
\item{string}{The string to trim.}
}
\value{
The revised vector.
}
\description{
\code{naz} converts NA to zero
}
\details{
Converts all NAs in a vector to zero.
}
\examples{
naz(NA)
}
\keyword{naz}

