% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btools_utilities.r
\name{odd}
\alias{odd}
\title{return odd elements of a vector}
\usage{
odd(x)
}
\arguments{
\item{x}{vector}
}
\value{
numeric vector
}
\description{
return odd elements of a vector
}
\examples{
odd(1:10)
}
