% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btools_stats.r
\name{p25}
\alias{p25}
\title{Compute the sample 25th percentile}
\usage{
p25(x, na.rm = FALSE)
}
\arguments{
\item{x}{a numeric vector containing the values whose 25th percentile is to be computed.}

\item{na.rm}{a logical value indicating whether NA values should be stripped before the computation proceeds.}
}
\value{
numeric
}
\description{
Compute the sample 25th percentile
}
\examples{
p25(1:100)
p25(c(1:10, NA, 11:100), na.rm=TRUE)
}
