% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_load.R
\name{package_load}
\alias{package_load}
\title{Install and load packages}
\usage{
package_load(..., quiet = TRUE)
}
\arguments{
\item{...}{Strings corresponding to names of packages}

\item{quiet}{Boolean, defaults to TRUE. If true, no warnings or messages
(including package start-up messages) will be displayed. If false, these
warnings will be displayed.}
}
\description{
`package_load` receives a list of packages, installs any packages which are
not currently installed, and loads the packages into the current environment
}
\examples{
package_load("tidyverse", "pracma", "magrittr")
}
