% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btools_utilities.r
\name{pany}
\alias{pany}
\title{Compute the sample value for a specific percentile, p.}
\usage{
pany(x, p, na.rm = FALSE)
}
\arguments{
\item{x}{a numeric vector containing the values whose p-th percentile is to be computed.}

\item{p}{the percentile.}

\item{na.rm}{a logical value indicating whether NA values should be stripped before the computation proceeds.}
}
\value{
numeric
}
\description{
Compute the sample value for a specific percentile, p.
}
\examples{
pany(1:100, .33)
pany(c(1:10, NA, 11:100), .33, na.rm=TRUE)
}
