% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivotCacheExtractor.R
\name{pivotCacheExtractor}
\alias{pivotCacheExtractor}
\title{pivotCacheExtractor}
\usage{
pivotCacheExtractor(file)
}
\arguments{
\item{file}{An excel file with pivot tables.}
}
\value{
a list of dataframes, one for each pivot table cache in the excel file.
}
\description{
Extract the hidden table that feeds excel pivot tables. Each pivot table is extracted from the excel file.
}
\examples{
pivotCacheExtractor(file = system.file("extdata/test_sheet.xlsx",
package = "pivotCacheExtractor",mustWork = TRUE))
}
