% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPCA3D.R
\name{plotPCA3D}
\alias{plotPCA3D}
\title{Plot DESeq2's PCA plotting with Plotly 3D scatterplot}
\usage{
plotPCA3D(object, intgroup = "condition", ntop = 500,
  returnData = FALSE)
}
\arguments{
\item{object}{a DESeqTransform object, with data in assay(x), produced for example by either rlog or varianceStabilizingTransformation.}

\item{intgroup}{interesting groups: a character vector of names in colData(x) to use for grouping}

\item{ntop}{number of top genes to use for principal components, selected by highest row variance}

\item{returnData}{should the function only return the data.frame of PC1, PC2 and PC3 with intgroup covariates for custom plotting (default is FALSE)}
}
\value{
An object created by plot_ly, which can be assigned and further customized.
}
\description{
The function will generate a plot_ly 3D scatter plot image for
a 3D exploration of the PCA.
}
