% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmean.R
\name{pmean}
\alias{pmean}
\title{Take generalized mean of x}
\usage{
pmean(..., p = 1, weight = NULL)
}
\arguments{
\item{...}{Vector to take generalized mean of}

\item{p}{Power for general mean}

\item{weight}{Vector containing weights}
}
\value{
generalized mean
}
\description{
Take generalized mean of x
}
