% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/btools_datadocumentation.r
\docType{data}
\name{qpopst}
\alias{qpopst}
\title{Dataset with quarterly state population}
\format{A data frame with 1 row per state per date (first day of quarter)}
\source{
Census Bureau, as prepared by Boyd
}
\usage{
qpopst
}
\description{
Dataset with postal abbreviation, date, value (thousands)
See popst for details of annual data.
Interpolated from annual to quarterly by Boyd.
}
\author{
Don Boyd, 2015-04-21
}

