% Generated by roxygen2 (4.0.2): do not edit by hand
\name{qtiledf}
\alias{qtiledf}
\title{Get quantiles for a vector, return as data frame}
\usage{
qtiledf(vec, probs)
}
\arguments{
\item{vec}{Numeric vector. No default.}

\item{probs}{Numeric vector of quantiles. Default is c(0, .1, .25, .5, .75, .9, 1).}
}
\value{
Data frame with columns as quantiles
}
\description{
\code{qtiledf} get trend, seasonal, remainder for a vector that has time-series data
}
\details{
Very little error checking.
Useful after dplyr's group_by, in do command, which requires data frame input.
}
\examples{
df \%>\% group_by(year) \%>\% do(qtiledf(.$marginpct, c(.1, .25, .5, .75, .9)))
}
\keyword{qtiledf}

