% Generated by roxygen2 (4.0.1): do not edit by hand
\name{qtr}
\alias{qtr}
\title{Get the quarter for a given date}
\usage{
qtr(dt)
}
\arguments{
\item{dt}{The date input. No default.}
}
\description{
\code{qtr} get the quarter for a given date
}
\details{
get the quarter for a given date
}
\examples{
qtr("2012-05-01")
}
\keyword{qtr}

