% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btools_dates.r
\name{qy}
\alias{qy}
\title{Create a date for start of quarter, from quarter, year}
\usage{
qy(q, y)
}
\arguments{
\item{q}{the quarter input. No default.}

\item{y}{the year input. No default.}
}
\description{
\code{qy} create a date for start of quarter, from quarter, year
}
\details{
You must ensure that the inputs are acceptable (e.g., quarter is in 1:4). The function does not check.
}
\examples{
qy(2, 1985)
}
\keyword{qy}
