% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcloneSMB.R
\name{rclone_obscure}
\alias{rclone_obscure}
\title{Obscure Password}
\usage{
rclone_obscure(pass, rclone_location = Sys.getenv("rclone_location"))
}
\arguments{
\item{pass}{Plain-text password}

\item{rclone_location}{Location of \code{rclone}}
}
\value{
Password obscured by rclone
}
\description{
rclone requires passwords to shares are obscured. This function
takes a plain-text password and obscures it using \verb{rclone obscure}.
}
\keyword{internal}
