% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/btools_datadocumentation.r
\docType{data}
\name{recessions}
\alias{recessions}
\title{Dataset with start and end dates of recessions}
\format{A data frame with 1 row per recession}
\source{
National Bureau of Economic Research (NBER)
http://www.nber.org/cycles.html
}
\usage{
recessions
}
\description{
Dataset with start and end dates of recessions, and first date of quarters in which recession started and ended
}
\author{
Don Boyd, 2015-04-21
}

