% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_blanks.R
\name{remove_blanks}
\alias{remove_blanks}
\title{Remove blanks from OTU table}
\usage{
remove_blanks(phylo, runID = "Run_number", blankID = "Group",
  blankName = "blank", removeBlank = F)
}
\arguments{
\item{phylo}{A phyloseq object with all samples per run.}

\item{runID}{The column name used to identify the different runs/plates.}

\item{blankID}{The column name used to identify the blanks.}

\item{blankName}{The name of the factor which describes the blanks.}

\item{removeBlank}{Should blank sample be removed from resulting phyloseq object.}
}
\value{
A phyloseq object with the OTU's removed from each plate per plate-blank
}
\description{
This function will remove the OTU from blanks within the
an OTU table, for each plate/run
}

