% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renameSMB.R
\name{renameSMB}
\alias{renameSMB}
\title{renameSMB}
\usage{
renameSMB(
  username,
  password,
  my_name = Sys.info()[["nodename"]],
  remote_name,
  service_name,
  old_path,
  new_path,
  domain,
  port = 445,
  timeout = 30,
  python_location
)
}
\arguments{
\item{username}{Your SMB user name}

\item{password}{Your SMB user password, do not store this in your code!}

\item{my_name}{The machine you are running this operation on (Default : \code{Sys.info()[["nodename"]]})}

\item{remote_name}{The smb server name}

\item{service_name}{The share name i.e. \code{'Share'}}

\item{old_path}{The complete path including file name and extension to your file e.g. \code{'Files/Reporting/test.txt'}}

\item{new_path}{The complete path including file name and extension to your new location e.g. \code{'Files/Reporting/test_renamed.txt'}}

\item{domain}{SMB server domain}

\item{port}{Common port  (Default : \code{445})}

\item{timeout}{How long to keep the connection open before timeout (Default : \code{30})}

\item{python_location}{The location of your python installation}
}
\value{
Returns no value
}
\description{
Use python SMBConnection to rename/move a file at path on a SMB share.
}
