% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieveFileSMB.R
\name{retrieveFileSMB}
\alias{retrieveFileSMB}
\title{retrieveFileSMB}
\usage{
retrieveFileSMB(username, password, my_name = Sys.info()[["nodename"]],
  remote_name, service_name, path, domain, file_obj, port = 445,
  timeout = 30, python_location)
}
\arguments{
\item{username}{Your SMB user name}

\item{password}{Your SMB user password, do not store this in your code!}

\item{my_name}{The machine you are running this operation on (Default : \code{Sys.info()[["nodename"]]})}

\item{remote_name}{The smb server name}

\item{service_name}{The service_name name i.e. \code{'Share'}}

\item{path}{The SMB folder path where your file is e.g. \code{'Files/Reporting/MyData.csv'}}

\item{domain}{SMB server domain}

\item{file_obj}{Full file path to put the retrieved data e.g. \code{'/home/username/file.txt'}}

\item{port}{Common port  (Default : \code{445})}

\item{timeout}{How long to keep the connection open before timeout (Default : \code{30})}

\item{python_location}{The location of your python installation}
}
\value{
Boolean indicating file exists at \code{file_obj}
}
\description{
Use python SMBConnection to retrieve the contents of a file at path on a SMB service_name. If the file already exists on the \code{file_obj}, it will be truncated and overwritten.
}
