% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btools_stats.r
\name{rollmean}
\alias{rollmean}
\title{Rolling mean}
\usage{
rollmean(x, nobs)
}
\arguments{
\item{x}{The vector to operate on}

\item{nobs}{Number of observations to include in window, including current}
}
\description{
\code{rollmean} rolling sample standard deviation
}
\details{
rolling mean
}
\examples{
rollmean(1:100, 4)
}
\keyword{rollmean}
