% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btools_stats.r
\name{rollmin}
\alias{rollmin}
\title{Rolling minimum}
\usage{
rollmin(x, nobs)
}
\arguments{
\item{x}{The vector to operate on}

\item{nobs}{Number of observations to include in window, including current}
}
\description{
\code{rollmin} rolling minimum
}
\details{
rolling minimum
}
\examples{
rollmean(1:100, 4)
}
\keyword{rollmin}
