% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btools_stats.r
\name{rollsd}
\alias{rollsd}
\title{Rolling sample standard deviation}
\usage{
rollsd(x, nobs)
}
\arguments{
\item{x}{The vector to operate on}

\item{nobs}{Number of observations to include in window, including current}
}
\description{
\code{rollsd} rolling sample standard deviation
}
\details{
rolling sample standard deviation
}
\examples{
rollsd(1:100, 4)
}
