% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btools_stats.r
\name{rollsd_p}
\alias{rollsd_p}
\title{Rolling populatoin standard deviation}
\usage{
rollsd_p(x, nobs)
}
\arguments{
\item{x}{The vector to operate on}

\item{nobs}{Number of observations to include in window, including current}
}
\description{
\code{rollsd_p} rolling population standard deviation
}
\details{
rolling population standard deviation
}
\examples{
rollsd_p(1:100, 4)
}
\keyword{rollsd_p}
