% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btools_time_series.r
\name{sa}
\alias{sa}
\title{Get seasonally adjusted values for a vector}
\usage{
sa(value, freq, s.window=9)
}
\arguments{
\item{value}{The vector to operate on.}

\item{freq}{Frequency of the time series (e.g., 1, 4, 12)}

\item{s.window}{Window parameter that is passed to stats::stl default=9}
}
\description{
\code{sa} Get seasonally adjusted values for a vector.
}
\details{
Seasonally adjusted version of value
}
\examples{
set.seed(1234)
n <- 28
x <- 1:n + rnorm(n)
x <- x * c(1.1, 1, .9, 1)
cbind(x, sa(x, 4))
}
\keyword{sa}
