% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btools_utilities.r
\name{safe.ifelse}
\alias{safe.ifelse}
\title{ifelse that can be used safely with dates}
\usage{
safe.ifelse(cond, yes, no)
}
\arguments{
\item{cond}{Logical expression.}

\item{yes}{Resulting value if cond is TRUE.}

\item{no}{Resulting value if cond is FALSE.}
}
\description{
ifelse that can be used safely with dates
}
\examples{
# snippet: mutate(date=safe.ifelse(freq=="A", as.Date(paste0(olddate, "-01-01")), date))
}
