% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safe_seq.R
\name{safe_seq}
\alias{safe_seq}
\title{Method of sequencing that throws an error if the sequence is not correct}
\usage{
safe_seq(start, end)
}
\arguments{
\item{start}{Starting point}

\item{end}{Ending point}
}
\value{
Sequence of integers
}
\description{
Method of sequencing that throws an error if the sequence is not correct
}
