% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveSMB.R
\name{saveSMB}
\alias{saveSMB}
\title{saveSMB}
\usage{
saveSMB(userID, password, client_machine_name = Sys.info()[["nodename"]],
  server_name, share, output_path, output_file_name, domain, file_location,
  port = 445, timeout = 30, python_location)
}
\arguments{
\item{userID}{Your SMB user name}

\item{password}{Your SMB user password, do not store this in your code!}

\item{client_machine_name}{The machine you are running this operation on (Default : \code{Sys.info()[["nodename"]]})}

\item{server_name}{The smb server name}

\item{share}{The share name i.e. \code{'Share'}}

\item{output_path}{The folder path to put your file e.g. \code{'Files/Reporting/'} to output the file to the Reporting subfolder of the Files root folder.}

\item{output_file_name}{The name and extension of the file you are saving e.g. \code{'MyData.csv'}""}

\item{domain}{SMB server domain}

\item{file_location}{Full file path e.g. \code{'\home\username\file.txt'}}

\item{port}{Common port  (Default : \code{445})}

\item{timeout}{How long to keep the connection open before timeout (Default : \code{30})}

\item{python_location}{The location of your python installation}
}
\value{
Returns no value
}
\description{
Use python SMBConnection to store the contents of a file at path on a SMB share. If the file already exists on the remote server, it will be truncated and overwritten.
}
