% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scientific_label.R
\name{scientific_label}
\alias{scientific_label}
\title{Properly display scientific scales for ggplot}
\usage{
scientific_label(x)
}
\arguments{
\item{x}{None}
}
\description{
Properly display scientific scales for ggplot
}
