% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_source.R
\name{script_source}
\alias{script_source}
\title{Source scripts into directory}
\usage{
script_source(..., directory = "R/", extension = ".R")
}
\arguments{
\item{...}{String vector containing scripts to be run. If
`extension = ".R"``, the names of the script are all that need to be
included.}

\item{directory}{Directory containins the scripts. Default directory is "R/".}

\item{extension}{File extension for scripts. Default extension is ".R".}
}
\description{
Source scripts into directory
}
\examples{
`script_source("function1", "function2")`
## Loads the functions "R/function1.R" and "R/function2.R"


}
