% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btools_utilities.r
\name{setdiff_all}
\alias{setdiff_all}
\title{which elements are not in the intersection of two vectors?}
\usage{
setdiff_all(v1, v2)
}
\arguments{
\item{v1}{vector}

\item{v2}{vector of the same type as v1}
}
\value{
a vector of items that are not in the intersection of two sets
}
\description{
which elements are not in the intersection of two vectors?
}
\examples{
v1 <- c(1, 3, 4, 5, 6, 7)
v2 <- c(2, 4, 6, 8, 10)
setdiff_all(v1, v2)
# compare to setdiff
setdiff(v1, v2)
setdiff(v2, v1)
}
