% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift.R
\name{shift}
\alias{shift}
\title{Shift columns of dataframe}
\usage{
shift(
  data = NULL,
  column = "",
  shift_by = 1,
  back = TRUE,
  order_by = NULL,
  fill_with = 0
)
}
\arguments{
\item{data}{Dataframe containing column to shift}

\item{column}{String, column to shift}

\item{shift_by}{Int, number of columns to shift by}

\item{back}{Boolean, if true shift back, if false shift forward}

\item{order_by}{String, column name to reorder prior to shifting}

\item{fill_with}{Float, fill shifted spaces with value}
}
\value{
Original dataframe with shifted values
}
\description{
Shift columns of dataframe
}
