% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/btools_datadocumentation.r
\docType{data}
\name{spi.a}
\alias{spi.a}
\title{Dataset with annual state personal income}
\format{A data frame with 1 row per state per year}
\source{
Bureau of Economic Analysis, via API, as prepared by Boyd
}
\usage{
spi.a
}
\description{
Dataset with postal abbreviation, year, value ($ thousands)
}
\author{
Don Boyd, 2015-04-25
}

