% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/btools_datadocumentation.r
\docType{data}
\name{spi.q}
\alias{spi.q}
\title{Dataset with quarterly state population}
\format{A data frame with 1 row per state per date (first day of quarter)}
\source{
Bureau of Economic Analysis, via API, as prepared by Boyd
}
\usage{
spi.q
}
\description{
Dataset with postal abbreviation, date, value ($ thousands)
}
\author{
Don Boyd, 2015-04-25
}

