% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/btools_datadocumentation.r
\docType{data}
\name{spop.a}
\alias{spop.a}
\title{Dataset with annual state population}
\format{A data frame with 1 row per state per year}
\source{
Census Bureau, as prepared by Boyd
}
\usage{
spop.a
}
\description{
Dataset with postal abbreviation, year, value (thousands), and esttype (estimate type)
Intercensal (July) estimates wherever possible, but only decennial was available in 1970, 1980
}
\author{
Don Boyd, 2015-04-25
}

