% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btools_states.r
\name{stabbr}
\alias{stabbr}
\title{get postal abbreviations given state names}
\usage{
stabbr(stname)
}
\arguments{
\item{stname}{character vector of state names}
}
\value{
character vector of state postal abbreviations
}
\description{
get postal abbreviations given state names
}
\examples{
stabbr(c("Alaska", "District of Columbia", "United States"))
}
