% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{stcodes}
\alias{stcodes}
\title{Dataset with state codes and names}
\format{A data frame with 1 row per state}
\source{
Census Bureau, other
}
\usage{
stcodes
}
\description{
Dataset with postal abbreviation, fips and Census codes, and state name
}
\author{
Don Boyd, 2014-05-27
}

