% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btools_dplyrtools.r
\name{stldf}
\alias{stldf}
\title{Get trend, seasonal, remainder for a vector that has time-series data}
\usage{
stldf(vec, freq)
}
\arguments{
\item{vec}{Numeric vector with time-series data. No default.}

\item{freq}{Frequency of the data. Numeric. Should be 4 (quarterly) or 12 (monthly). No default.}
}
\value{
Data frame with 3 columns: trend, seasonal, remainder
}
\description{
\code{stldf} get trend, seasonal, remainder for a vector that has time-series data
}
\details{
Returns a data frame with 3 columns: trend, seasonal, remainder. Very little error checking.
Useful after dplyr's group_by, in do command, which requires data frame input. Make sure data are sorted by time before using.
}
\examples{
library(bdata) # so that spop.q is available
library(dplyr)
spop.q \%>\% group_by(stabbr) \%>\%
    arrange(date) \%>\% # BE SURE DATA HAVE BEEN SORTED BY DATE WITHIN GROUPING VARS!!!
    do(cbind(., stldf(.$value, 4)))
}
\keyword{stldf}
