% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btools_states.r
\name{stname}
\alias{stname}
\title{get state names given postal abbreviations}
\usage{
stname(stabbr)
}
\arguments{
\item{stabbr}{character vector of state postal abbreviations}
}
\value{
character vector of state names
}
\description{
get state names given postal abbreviations
}
\examples{
stname(c("AK", "DC", "US"))
}
