% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storeFileSMB.R
\name{storeFileSMB}
\alias{storeFileSMB}
\title{storeFileSMB}
\usage{
storeFileSMB(
  username,
  password,
  my_name = Sys.info()[["nodename"]],
  remote_name,
  service_name,
  path,
  domain,
  file_obj,
  port = 445,
  timeout = 30,
  python_location
)
}
\arguments{
\item{username}{Your SMB user name}

\item{password}{Your SMB user password, do not store this in your code!}

\item{my_name}{The machine you are running this operation on (Default : \code{Sys.info()[["nodename"]]})}

\item{remote_name}{The smb server name}

\item{service_name}{The service_name name i.e. \code{'Share'}}

\item{path}{The path to put your file e.g. \code{'/home/username/file.txt'}}

\item{domain}{SMB server domain}

\item{file_obj}{Full file path, to be read e.g. \code{'Files/Reporting/MyData.csv'}}

\item{port}{Common port  (Default : \code{445})}

\item{timeout}{How long to keep the connection open before timeout (Default : \code{30})}

\item{python_location}{The location of your python installation}
}
\value{
Returns no value
}
\description{
Use python SMBConnection to store the contents of a file at path on a SMB service_name. If the file already exists on the remote server, it will be truncated and overwritten.
}
