% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btools_strings.r
\name{str_detect_any}
\alias{str_detect_any}
\title{Detect whether any vector elements are in a string}
\usage{
str_detect_any(s, elements)
}
\arguments{
\item{s}{A string.}

\item{elements}{Vector elements to look for.}
}
\value{
Logical indicating whether any of \code{elements} are found in \code{s}.
}
\description{
Detect whether any vector elements are in a string
}
\examples{
str_detect_any("abc defg ijk", c("123", "def", "11"))
str_detect_any("abc defg ijk", c("123", "xyz", "11"))
}
