% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btools_utilities.r
\name{str_extract_after_last}
\alias{str_extract_after_last}
\title{Return portion of a string after last occurrence of a pattern}
\usage{
str_extract_after_last(s, last)
}
\arguments{
\item{s}{A string.}

\item{last}{Pattern to look for.}
}
\value{
Substring of \code{s} after last occurrence of \code{last}.
}
\description{
Return portion of a string after last occurrence of a pattern
}
\examples{
str_extract_after_last("ab#cde#yzy", "#")
str_extract_after_last("abcde#yzy", "c")
}
