% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btools_strings.r
\name{str_extract_before_first}
\alias{str_extract_before_first}
\title{Return portion of a string before first occurrence of a pattern}
\usage{
str_extract_before_first(s, first)
}
\arguments{
\item{s}{A string.}

\item{first}{Pattern to look for.}
}
\value{
Substring of \code{s} before first occurrence of \code{first}.
}
\description{
Return portion of a string before first occurrence of a pattern
}
\examples{
str_extract_before_first("abcde#yzy", "#")
str_extract_before_first("abcde#yzy", "y")
}
