% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btools_utilities.r
\name{trim.ws}
\alias{trim.ws}
\title{Trim white space at either end of strings}
\usage{
trim.ws(s)
}
\arguments{
\item{s}{The string to trim.}
}
\value{
The trimmed string.
}
\description{
\code{trim.ws} trims white space around strings
}
\details{
All white space is removed from the ends.
}
\examples{
trim.ws("   original string has leading and trailing spaces   ")
}
\keyword{trim.ws}
