% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btools_utilities.r
\name{trimlead}
\alias{trimlead}
\title{Trim leading white space in strings}
\usage{
trimlead(s)
}
\arguments{
\item{s}{The string to trim.}
}
\value{
The trimmed string.
}
\description{
\code{trimlead} trims leading white space in strings
}
\details{
All white space is removed from the leading (left) side of the string.
}
\examples{
trimlead("   original string has leading and trailing spaces   ")
}
\keyword{trimlead}
