% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btools_utilities.r
\name{trimtrail}
\alias{trimtrail}
\title{Trim trailing white space in strings}
\usage{
trimtrail(s)
}
\arguments{
\item{s}{The string to trim.}
}
\value{
The trimmed string.
}
\description{
\code{trimtrail} trims trailing white space in strings
}
\details{
All white space is removed from the trailing (right) side of the string.
}
\examples{
trimtrail("   original string has leading and trailing spaces   ")
}
\keyword{trimtrail}
