% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_btools.R
\name{update_btools}
\alias{update_btools}
\title{Install pmder package}
\usage{
update_btools()
}
\description{
Execution of this function will install the latest version of the `btools` package into the system.
}
