% Generated by roxygen2 (4.0.2): do not edit by hand
\name{uvf}
\alias{uvf}
\title{Get expression based on quoted variables names, to use in dplyr's select clause}
\usage{
uvf(vars)
}
\arguments{
\item{vars}{The character vector of variable names (column names in a data frame)}
}
\description{
\code{uvf} get expression based on quoted variables names, to use in dplyr's select clause
}
\details{
get expression based on quoted variables names, to use in dplyr's select clause
}
\examples{
# use uvf within the select clause in the dplyr package
library(dplyr)
library(btools)
df <- data.frame(a=1:4, b=4:1, c=7:10, d=9:12)
vars <- c("c", "d")
df \%>\% select(a, eval(uvf(vars)))
}
\keyword{uvf}

