% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btools_utilities.r
\name{xdate}
\alias{xdate}
\title{Convert Excel numeric date to date.}
\usage{
xdate(xdate)
}
\arguments{
\item{xdate}{a numeric vector containing dates in Excel format.}
}
\value{
date
}
\description{
Convert Excel numeric date to date.
}
\examples{
xdate(30000)
}
