% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btools_utilities.r
\name{xldate}
\alias{xldate}
\title{Convert Excel numeric date to date.}
\usage{
xldate(xldate)
}
\arguments{
\item{xldate}{a numeric vector containing dates in Excel format.}
}
\value{
date
}
\description{
Convert Excel numeric date to date.
}
\examples{
xldate(30000)
}
