% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsxPivotCacheExtractor.R
\name{xlsxPivotCacheExtractor}
\alias{xlsxPivotCacheExtractor}
\title{xlsxPivotCacheExtractor}
\usage{
xlsxPivotCacheExtractor(file)
}
\arguments{
\item{file}{An excel xlsx file with pivot tables.}
}
\value{
a list of dataframes, one for each pivot table cache in the excel file.
}
\description{
Extract the hidden table that feeds excel pivot tables. Each pivot table is extracted from the excel file.
}
\examples{
xlsxPivotCacheExtractor(file = system.file("extdata/test_sheet.xlsx",
package = "btools",mustWork = TRUE))
}
