% Generated by roxygen2 (4.0.0): do not edit by hand
\name{year}
\alias{year}
\title{Extract year from a date}
\usage{
year(dt)
}
\arguments{
  \item{dt}{The date input. No default.}
}
\description{
\code{year} extract year from a date
}
\details{
extract year from a date
}
\examples{
year("2012-05-01")
}
\keyword{year}

