% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btools_dates.r
\name{yq2}
\alias{yq2}
\title{First day of quarter, from year, quarter}
\usage{
yq2(y, q)
}
\arguments{
\item{y}{the year input. No default.}

\item{q}{the quarter input. No default.}
}
\description{
\code{yq2} first day quarter, with separate year, quarter inputs
}
\examples{
yq2(2020, 1)
yq2("2020", "1")
}
