% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/btools_utilities.r
\name{zpad}
\alias{zpad}
\title{Zero-pad leading characters in a numeric value.}
\usage{
zpad(vec)
}
\arguments{
\item{outlen}{The length of the string to be returned. Default is 2.}

\item{vec}{The object to zero-pad.}
}
\value{
Zero-padded strings of length outlen.
}
\description{
\code{zpad} converts NA to zero
}
\details{
Zero-pads on the left all numeric in a vector, returning strings of length outlen.
}
\examples{
zpad(1, 4)
}
\keyword{zpad}

