% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_contributions.R
\name{analyze_contributions}
\alias{analyze_contributions}
\title{Add meta data to contributions file}
\usage{
analyze_contributions(contributions_fp, mappingfile_fp)
}
\arguments{
\item{contributions_fp}{File location of the metagenomic_contributions.py output.}

\item{mappingfile_fp}{File location of the input sample metadata.}
}
\value{
A very long dataframe with taxa information and sample metadata for each observed OTU.
}
\description{
This function loads a file as a matrix. It assumes that the first column
contains the rownames and the subsequent columns are the sample identifiers.
Any rows with duplicated row names will be dropped with the first one being
kepted.
}

