% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btools-package.R
\docType{package}
\name{btools}
\alias{btools-package}
\alias{btools}
\title{btools: General text and data management tools}
\description{
The `btools` package includes utilities for working with strings, dates, and 
numeric data, and utilities for dplyr.
}
\author{
\strong{Maintainer}: Don Boyd \email{donboyd5@gmail.com}

}
