% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/btools_utilities.r
\name{capwords}
\alias{capwords}
\title{Capitalize first letter of each word}
\usage{
capwords(s)
}
\arguments{
\item{s}{The string to capitalize words of}
}
\value{
The initial-capped result.
}
\description{
\code{capwords} capitalize first letter of each word
}
\details{
All white space is removed from the trailing (right) side of the string.
}
\examples{
capwords("string to capitalize words in")
}
\keyword{capwords}

