% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_alpha_diversity.R
\name{compare_alpha_diversity}
\alias{compare_alpha_diversity}
\title{Calculate alpha diversity statistics}
\usage{
compare_alpha_diversity(physeq, x = "Day", group = "Treatment",
  diversity = c("Observed", "Shannon", "Simpson"),
  test_type = c("nonparametric", "parametric"),
  col_var = "PD_whole_tree_alpha", num_perm = 999,
  multiple_corrections = T, write = F, filename = "results", ...)
}
\arguments{
\item{metagenomic_contributions.py}{output file}

\item{mapping}{file with sample metadata}

\item{list}{of column variables to add to new table}
}
\value{
A dataframe with taxa information and sample metadata
}
\description{
Compute p-values and multiple comparisons adjusted q-values for
two-group comparisons across multiple timepoints.
}

