% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_categories.R
\name{compare_beta_diversity}
\alias{compare_beta_diversity}
\title{Calculate alpha diversity statistics}
\usage{
compare_beta_diversity(phylo, x = "Day", group = "Treatment",
  test = "adonis", bdiv = "unweighted", write = F, filename = "results",
  ...)
}
\arguments{
\item{phylo}{phyloseq object}
}
\value{
A dataframe for an PERMANOVA test over each timepoint from each two group comparison.
}
\description{
Compute p-values and multiple comparisons adjusted q-values for
two-group comparisons across multiple timepoints.
}

