% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_beta_diversity.R
\name{compare_beta_diversity}
\alias{compare_beta_diversity}
\title{Calculate beta diversity statistics over time.}
\usage{
compare_beta_diversity(phylo, x = as.character(),
  group = as.character(), test = c("adonis", "anosim"),
  bdiv = c("weighted", "unweighted"), write = F,
  filename = "results", fdr = T, fdr_test = "fdr", seed = 918, ...)
}
\arguments{
\item{phylo}{A phyloseq object}

\item{x}{The variable that describes Time}

\item{group}{The variable that describes 2 or more groups}

\item{test}{The PERMANOVA test to use.}

\item{bdiv}{Beta diversity metric to calculate significance.}

\item{write}{Should results be written to a file in the current working directory.}

\item{filename}{The name of the output file if write is TRUE.}

\item{fdr}{Should FDR correction be applied.}

\item{fdr_test}{The test to use to correct for multiple comparisons. Default is FDR.}

\item{seed}{The seed number to use when calculating beta diversity metrics. Default is 918.}

\item{...}{Any additional arguments for external functions.}
}
\value{
A dataframe for an PERMANOVA test over each timepoint from each two group comparison.
}
\description{
Compute p-values and multiple comparisons adjusted q-values for
two-group comparisons using PERMANOVA on beta diversity metrics across multiple timepoints.
}
