% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btools_utilities.r
\name{cton}
\alias{cton}
\title{Convert character to numeric}
\usage{
cton(cvar)
}
\arguments{
\item{cvar}{The character string input. No default.}
}
\description{
\code{cton} converts character to numeric
}
\details{
Replaces spaces, comma, $, and percent sign in a string with NULL and then converts to numeric.
Keeps letters so that scientific notation will be evaluated properly.
Also look at \code{extract_numeric} in package stringr
}
\examples{
char <- "$198,234.75"
cton(char)
}
\keyword{cton}
