% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btools_dates.r
\name{fdoq}
\alias{fdoq}
\title{First day of quarter}
\usage{
fdoq(date)
}
\arguments{
\item{date}{a vector of dates. No default.}
}
\description{
\code{fdoq} create a date for start of quarter, from quarter, year
}
\details{
You must ensure that the inputs are acceptable (e.g., the input data are dates). The function does not check.
}
\examples{
fdoq(as.Date("2010-01-03"))
x <- as.Date(c("2010-01-03", "2011-03-03", "2011-04-07"))
fdoq(x)
}
\keyword{fdoq}
