% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ht}
\alias{ht}
\title{Show head and tail of a vector, matrix, table, data frame or function}
\usage{
ht(obj)
}
\arguments{
\item{nrecs}{number of records, rows, whatever to show at head and at tail}

\item{obj}{The object. No default.}
}
\description{
\code{ht} head and tail of a vector, matrix, table, data frame or function
}
\details{
show head and tail of a vector, matrix, table, data frame or function
}
\examples{
ht(recessions, 4)
}
\keyword{ht}

