% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_rcc.R
\name{import_rcc}
\alias{import_rcc}
\title{Import .RCC files}
\usage{
import_rcc(rccPath = "", useFolderPath = T)
}
\arguments{
\item{RCC}{The location of the raw nanostring files. Input should only be the
location of a folder in quotes.}
}
\value{
A dataframe of count data with columns as samples and rows as genes.
}
\description{
\code{import_rcc} Import .RCC files from NanoString analysis and
convert into an easily read table. Headers will be sample ID's. Adapted from NanoStringNorm.
}

