% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btools_dates.r
\name{ldom}
\alias{ldom}
\title{Last day of month}
\usage{
ldom(date)
}
\arguments{
\item{date}{a vector of dates. No default.}
}
\description{
\code{ldom} get the last day of month, for a given date
}
\details{
You must ensure that the inputs are acceptable (e.g., the input data are dates). The function does not check.
}
\examples{
ldom(as.Date("2010-01-03"))
x <- as.Date(c("2010-01-01", "2010-01-31","2019-02-01", "2020-02-01"))
ldom(x)
}
\keyword{ldom}
