% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btools_utilities.r
\name{ma4}
\alias{ma4}
\title{Get 4-period moving average (3 lags + current)}
\usage{
ma4(x)
}
\arguments{
\item{x}{The vector to operate on.}
}
\description{
\code{ma4} get 4-period moving average
}
\details{
4-period moving average
}
\examples{
ma4(7:21)
}
\keyword{ma4}
