% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btools_utilities.r
\name{naz}
\alias{naz}
\title{Convert NA to zero}
\usage{
naz(vec)
}
\arguments{
\item{vec}{The vector to convert.}
}
\value{
The revised vector.
}
\description{
\code{naz} converts NA to zero
}
\details{
Converts all NAs in a vector to zero.
}
\examples{
naz(NA)
}
\keyword{naz}
