% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btools_utilities.r
\name{ns}
\alias{ns}
\title{create vector of sorted names of an object}
\usage{
ns(obj)
}
\arguments{
\item{obj}{named object}
}
\value{
a vector of sorted names
}
\description{
create vector of sorted names of an object
}
\examples{
names(iris) # unsorted
ns(iris)
}
