% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btools_dplyrtools.r
\name{qtiledf}
\alias{qtiledf}
\title{Get quantiles and number of not-NA observations for a vector, return as tibble}
\usage{
qtiledf(vec, probs)
}
\arguments{
\item{vec}{Numeric vector. No default.}

\item{probs}{Numeric vector of quantiles. Default is c(0, .1, .25, .5, .75, .9, 1).}
}
\value{
Data frame with columns as quantiles
}
\description{
\code{qtiledf} get quantiles and number of not-NA observations for a vector, return as data frame
}
\details{
Very little error checking.
Useful after dplyr's group_by, in do command, which requires data frame input.
}
\examples{
library(dplyr)
df <- tibble(year=c(rep(1, 5), rep(2, 7)), x=c(seq(1, 2, length.out=11), NA))
df
df \%>\%
  group_by(year) \%>\%
  summarise(qtiledf(.$x, c(.1, .25, .5, .75, .9)))
}
\keyword{qtiledf}
