% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{data}
\name{recessions}
\alias{recessions}
\title{Dataset with start and end dates of recessions}
\format{A data frame with 1 row per recession}
\source{
National Bureau of Economic Research (NBER)
}
\usage{
recessions
}
\description{
Dataset with start and end dates of recessions, and first date of quarters in which recession started and ended
}
\author{
Don Boyd, 2014-05-27
}

