% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/btools_datadocumentation.r
\docType{data}
\name{stcodes}
\alias{stcodes}
\title{Dataset with state codes and names}
\format{A data frame with 1 row per state}
\source{
Census Bureau, other
}
\usage{
stcodes
}
\description{
Dataset with postal abbreviation, fips and Census codes, and state name
}
\author{
Don Boyd, 2015-01-06
}

