% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btools_utilities.r
\name{sum4}
\alias{sum4}
\title{Get 4-period moving sum (3 lags + current)}
\usage{
sum4(x)
}
\arguments{
\item{x}{The vector to operate on.}
}
\description{
\code{sum4} get 4-period moving sum
}
\details{
4-period moving sum
}
\examples{
sum4(7:21)
}
\keyword{sum4}
