% Generated by roxygen2 (4.0.1): do not edit by hand
\name{trimws}
\alias{trimws}
\title{Trim white space at either end of strings}
\usage{
twimws(s)
}
\arguments{
\item{s}{The string to trim.}
}
\value{
The trimmed string.
}
\description{
\code{trimws} trims white space around strings
}
\details{
All white space is removed from the ends.
}
\examples{
trimws("   original string has leading and trailing spaces   ")
}
\keyword{trimws}

